+++ src/modules/serverProviders.module.ts
@@ -1,0 +1,28 @@
+import {NgModule, ModuleWithProviders} from '@angular/core';
+
+/**
+ * Module containing all standard server providers
+ */
+@NgModule(
+{
+    imports: [],
+    providers: []
+})
+export class ServerProvidersModule
+{
+    //######################### public methods #########################
+    
+    /**
+     * Returns module with all server providers
+     */
+    public static forRoot(): ModuleWithProviders 
+    {
+        return {
+            ngModule: ServerProvidersModule,
+            providers: 
+            [
+            ]
+        };
+    }
+
+}